requirements.txt
selenium==4.15.2
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0